puzzle
working out my ordering. SDL3_vulkan gets us to Dispose for a trip to VulkanCore
Dispose is Boss of this one and will create and destroy every handle like RAII cookie monster.
Dispose heart, VulkanCore is Spine. Rest is trash :D
Dispose is also BufferManager <3
VulkanCore is going to make BLAS and TLAS for the eyeballs.